requirements.txt
# base Streamlit + data libraries
streamlit==1.45.1
pandas==2.2.3
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
streamlit-aggrid==1.1.4.post1
xlsxwriter==3.2.3
openpyxl==3.1.5
python-dotenv==1.1.0

# LangChain ecosystem (0.3.x)
langchain==0.3.25
langchain-community==0.3.24
langchain-experimental

# Google Vertex AI / Gemini integration
langchain-google-vertexai==2.0.24
#google-cloud-aiplatform>=1.38.0
google-cloud-aiplatform
vertexai 
pyarrow
qdrant-client